[OMCSession] split file
